[Feat] 스킬 효과 네펜데스에 적용

+++ Assets/WorkSpace/LJH/Scripts/Dungeon/Monster/BT/Condition Node/IsDieNode.cs
@@ -2,9 +2,9 @@ using System;
 
 public class IsDieNode : BTNode
 {
-    private Func<int> getHp;
+    private Func<float> getHp;
 
-    public IsDieNode(Func<int> getHp)
+    public IsDieNode(Func<float> getHp)
     {
         this.getHp = getHp;
     }
