+++ Assets/WorkSpace/LJH/Scripts/Dungeon/Monster/MonsterModel.cs
@@ -6,8 +6,8 @@ public class MonsterModel : MonoBehaviour
 
     public int Id;
     public string Name;
-    public int MaxHp;
-    public int Hp;
+    public float MaxHp;
+    public float Hp;
     public int Attack;
     public float AttackRange;
     public float ChaseRange;
@@ -37,11 +37,11 @@ public class MonsterModel : MonoBehaviour
         AttackCooldown = monsterData.AttackCooldown;
 
     }
-    public int GetMonsterMaxHp() => MaxHp;
+    public float GetMonsterMaxHp() => MaxHp;
 
-    public int GetMonsterHp() => Hp;
+    public float GetMonsterHp() => Hp;
 
-    public void SetMonsterHp(int hp)
+    public void SetMonsterHp(float hp)
     {
         Hp += hp;
     }
